requirements.txt
fastapi<0.69.0
uvicorn<0.16.0
python-multipart
youtube-transcript-api
pydantic<2.0.0
jinja2<3.1.0
python-jose
python-dotenv
requests
fpdf2
markdown2
anthropic<1.0.0
httpx<0.25.0